requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Bcrypt>=1.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0

# Configuration management
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.0

# Image processing and face recognition
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
# TensorFlow for Python 3.12 - use latest compatible version
tensorflow>=2.15.0
h5py>=3.9.0
matplotlib>=3.7.0

# File handling and security
Werkzeug>=2.3.0

# Report generation
openpyxl>=3.1.0
pandas>=2.0.0
reportlab>=4.0.0

# HTTP requests
requests>=2.31.0

# System monitoring
psutil>=5.9.0

# Testing (for future use)
pytest>=7.4.0
pytest-mock>=3.11.0
hypothesis>=6.87.0

# Development tools
xlsxwriter>=3.1.0

# Machine learning
scikit-learn>=1.3.0
Flask-HTTPAuth>=4.8.0